minor refactors for verbiage and packge organization
