+++ docs/resources/system_subnet_pool_member.md
@@ -3,18 +3,24 @@
 page_title: "oxide_system_subnet_pool_member Resource - terraform-provider-oxide"
 subcategory: ""
 description: |-
-  This resource manages a subnet pool member using the system API.
+  This resource manages a subnet pool member.
 ---
 
 # oxide_system_subnet_pool_member (Resource)
 
-This resource manages a subnet pool member using the system API.
+This resource manages a subnet pool member.
 
 ## Example Usage
 
 ```terraform
+resource "oxide_system_subnet_pool" "example" {
+  name        = "my-subnet-pool"
+  description = "Example system subnet pool."
+  ip_version  = "v4"
+}
+
 resource "oxide_system_subnet_pool_member" "example" {
-  subnet_pool_id    = "3e2c6e84-bed8-4c94-afc3-1032082d6a90"
+  subnet_pool_id    = oxide_system_subnet_pool.example.id
   subnet            = "192.0.2.0/24"
   min_prefix_length = 24
   max_prefix_length = 28
